Time: 2127 ms (5.88%), Space: 43.5 MB (100.00%) - LeetHub

+++ 1162-as-far-from-land-as-possible/1162-as-far-from-land-as-possible.kt
@@ -1,26 +1,26 @@
 class Solution {
     fun maxDistance(grid: Array<IntArray>): Int {
-        var ans = -1L
+        var ans = -1
         for (i in grid.indices) {
             for (j in grid.indices) {
                 if(grid[i][j] == 1) {
                     continue
                 }
-                var temporaryMinDist = Int.MAX_VALUE.toLong()
+                var temporaryMinDist = Int.MAX_VALUE
                 for (k in grid.indices) {
                     for(l in grid.indices) {
                         if(grid[k][l] == 0) {
                             continue
                         }
-                        val currentDist = kotlin.math.abs(k - i).toLong() + kotlin.math.abs(l - j).toLong()
+                        val currentDist = kotlin.math.abs(k - i) + kotlin.math.abs(l - j)
                         temporaryMinDist = kotlin.math.min(temporaryMinDist, currentDist)
                     }
                 }
-                if(temporaryMinDist == Int.MAX_VALUE.toLong()) continue
+                if(temporaryMinDist == Int.MAX_VALUE) continue
                 ans = kotlin.math.max(temporaryMinDist, ans)
             }
         }
 
-        return ans.toInt()
+        return ans
     }
 }
